# src/main/resources/swagger/FileService.yml
swagger: '2.0'
host: 'localhost:8080'
basePath: /

info:
  title: APIs for File Operations
  description: APIs to deals with files to create, download and delete the file
  version: '1.0'
  contact:
    email: prashantenc@gmail.com

schemes: 
- "http"
- "https"

paths:
  '/file/create':
    post:
      tags:
        - Create the file
      summary: Return the status of create the file
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: file
          in: formData
          description: file
          required: true
          type: file
      responses:
        '200':
          description: Ok
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/file/download':
    get:
      tags:
        - Download File
      summary: Return the status of download the file
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: fileName
          in: query
          description: fileName
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/file/delete':
    delete:
      tags:
        - Delete File
      summary: Return the status of delete the file
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: fileName
          in: query
          description: fileName
          required: true
          type: string
      responses:
        '200':
          description: Ok
          schema:
            type: string
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden